Build the GNU tools using ecc.
git-svn-id: http://ellcc.org/svn/ellcc/trunk@3496 82857169-6e38-49d7-aaa8-07fb558f12fc

+++ gnu/src/gdb/bfd/elf32-v850.c
@@ -3142,7 +3142,7 @@ v850_elf_relax_section (bfd *abfd,
 						toaddr, align_pad_size))
 		goto error_return;
 
-	      for (i  = BFD_ALIGN (toaddr - align_pad_size, 1);
+	      for (i  = (int)BFD_ALIGN (toaddr - align_pad_size, 1);
 		   (i + 1) < toaddr; i += 2)
 		bfd_put_16 (abfd, NOP_OPCODE, contents + i);
